backend/pyproject.toml
[tool.poetry]
name = "ai-voice-banking-backend"
version = "0.1.0"
description = "FastAPI backend with ML and WebSocket layers for the AI Voice Assistant banking flow."
authors = ["Assistant"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-multipart = "^0.0.9"
httpx = "^0.27.0"
pydantic-settings = "^2.4.0"
redis = "^5.0.7"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
motor = "^3.4.0"
transformers = "^4.57.1"
torch = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.7"
black = "^24.4.2"
ruff = "^0.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

